requirements/base.txt
# Django 핵심
Django==4.2.7
python-dotenv==1.0.0

# 데이터베이스
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
djangorestframework-camel-case==1.4.2
django-filter==23.4
pymongo==4.6.3

# API 문서화
drf-spectacular==0.27.0

# CORS
django-cors-headers==4.3.1

# 보안
cryptography==45.0.5

# 이미지 처리
Pillow==10.1.0

# AWS SDK (모든 환경에서 필요)
boto3==1.34.0
botocore==1.34.0

# 파일 저장소 (S3)
django-storages==1.14.2
